usfhashmap, usfio and usfstring rework

+++ include/usfio.h
@@ -2,14 +2,15 @@
 #define USFIO_H
 
 #include <stdio.h>
+#include "usfstd.h"
 #include "usfstring.h"
 
-char *usf_ftos(char *file, uint64_t *l);
-void usf_printtxt(char **text, uint64_t len, FILE *stream);
-char **usf_ftot(char *file, uint64_t *l);
-char **usf_ftost(char *file, uint64_t *l);
-void usf_freetxt(char **text, uint64_t nlines);
-size_t usf_btof(char *file, void *pointer, size_t size);
-void *usf_ftob(char *file, size_t *size);
+char *usf_ftos(char *file, u64 *l);
+void usf_printtxt(char **text, u64 len, FILE *stream);
+char **usf_ftot(char *file, u64 *l);
+char **usf_ftost(char *file, u64 *l);
+void usf_freetxt(char **text, u64 nlines);
+u64 usf_btof(char *file, void *pointer, u64 size);
+void *usf_ftob(char *file, u64 *size);
 
 #endif
